Added recipe generator and idempotence help.
This commit adds 2 additional bits of functionality.

First, it provides a simple idempotence solution.
I've introduced a sunzi.run_recipe method which checks for the
existence of a recipe file in /etc/sunzi/completed_recipes on the
machine being provisioned. The existence of this file indicates
that the recipe has already been run. A recipe file is created
in this directory upon the initial run for each recipe.

Second, it provides a recipe file generator.
This CLI helper helps keep projects organized over time because
server provisioning is very much an incremental and evolving process.
The generator simply creates empty files under sunzi/recipes in the
following format TIMESTAMP-ROLE-RECIPE_NAME.sh